unittest work

+++ src/BitStream.h
@@ -16,10 +16,10 @@ public:
     unsigned int Pop(const unsigned int bits);
     bool TryPop(unsigned int& data, unsigned int bits);
     
-    void FlushFront();
-    void FlushBack();
+    void FlushFront(); // make sure the buffer starts at a whole byte
+    void FlushBack(); // fill the buffer with 0 so it ends on a whole byte
 
-    unsigned int BitsAvailable() const;
+    size_t BitsAvailable() const;
 
 private:
     std::deque<unsigned char> m_bigBuffer;
